[BTAT-250] Refctored Accounting Period Model, created Property Details model and Removed unused ConnectorResponseModel
