+++ app/models/AccountingPeriodModel.scala
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2017 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package models
+
+import java.time.LocalDate
+
+import play.api.libs.json.Json
+
+
+case class AccountingPeriodModel(start: LocalDate, end: LocalDate) {
+  val determineTaxYear = end match {
+    case x if x.getDayOfMonth > 5 && x.getMonthValue >= 4 => x.getYear + 1
+    case x => x.getYear
+  }
+}
+
+object AccountingPeriodModel {
+  implicit val format = Json.format[AccountingPeriodModel]
+}
