+++ app/models/BusinessListResponseModel.scala
@@ -40,16 +40,6 @@ case class BusinessModel(
                           businessPostcode: Option[String]
                           )
 
-case class AccountingPeriodModel(start: LocalDate, end: LocalDate) {
-  val determineTaxYear = end match {
-    case x if x.getDayOfMonth > 5 && x.getMonthValue >= 4 => x.getYear + 1
-    case x => x.getYear
-  }
-}
-
-object AccountingPeriodModel {
-  implicit val format = Json.format[AccountingPeriodModel]
-}
 
 object BusinessModel {
   implicit val format = Json.format[BusinessModel]
