+++ test/models/BusinessListResponseModelSpec.scala
@@ -75,22 +75,4 @@ class BusinessListResponseModelSpec extends UnitSpec with Matchers {
       Json.parse(businessErrorString).as[BusinessDetailsErrorModel] shouldBe businessErrorModel
     }
   }
-
-  "The AccountingPeriodModel Model" when {
-    "the end date is before the Start of the next Tax Year" should {
-      "return the current Tax Year" in {
-        AccountingPeriodModel("2017-04-06", "2018-04-05").determineTaxYear shouldBe 2018
-      }
-    }
-    "the end date is on the Start of the next Tax Year" should {
-      "return the next Tax Year" in {
-        AccountingPeriodModel("2017-04-07", "2018-04-06").determineTaxYear shouldBe 2019
-      }
-    }
-    "the end date is after the Start of the next Tax Year" should {
-      "return the next Tax Year" in {
-        AccountingPeriodModel("2017-04-08", "2018-04-07").determineTaxYear shouldBe 2019
-      }
-    }
-  }
 }
